pyproject.toml
[project]
name = "anantya-foundation-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.133.1",
    "firebase-admin>=7.2.0",
    "gspread>=6.2.1",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "resend>=2.23.0",
    "sendgrid>=6.12.5",
    "sib-api-v3-sdk>=7.6.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
]
